requirements.txt
pyquery
python-pushover